fix(loss): compute causal-LM cross-entropy in float32
CLMCrossEntropyLoss passed the model's logits straight into CrossEntropyLoss.
Under mixed precision those are bfloat16, and nothing promotes them: FSDP2's
MixedPrecisionPolicy casts parameters, it does not install torch.autocast, so
the operator-level float32 autocast list never applies. The log-softmax, its
backward, and -- because the reduction is "mean" -- the accumulation of the
per-token losses therefore all ran in bfloat16 over a 131k-entry vocabulary.

Both reference implementations up-cast at exactly this point:
TorchTitan does it on every one of its cross-entropy paths
(torchtitan/components/loss.py, e.g. `pred.flatten(0, 1).float()`), and HF
transformers opens ForCausalLMLoss with `logits = logits.float()`
(transformers/loss/loss_utils.py).

Measured against a float64 evaluation of identical bfloat16-quantized logits at
vocab 131072, the loss value was off by ~4.9e-2 and is now off by ~2.4e-6. The
accumulator dominates that figure; the log-softmax alone accounts for ~3e-3.
On a real 8B checkpoint and 262,144 held-out tokens the gradient into lm_head
differs by 0.19% between the two paths, and the up-cast recovers about 1.8x of
that -- the remainder is imposed by the bfloat16 logit boundary itself and is
not reachable from the loss.

After the change the loss and its gradient are bit-identical to TorchTitan's
formulation for bfloat16, float16 and float32 inputs at vocab 1024 and 131072.

Tests cover the three symptoms: the returned dtype, accuracy against a float64
reference, and agreement with TorchTitan's sum-reduction-over-valid-tokens
formulation. All three fail without the fix.

+++ src/modalities/loss_functions.py
@@ -48,7 +48,13 @@ class CLMCrossEntropyLoss(Loss):
         shift_logits = lm_logits.contiguous()
         shift_labels = labels.contiguous().long()
         # Flatten the tokens. We compute here, the loss per token.
-        loss = self.loss_fun(shift_logits.view(-1, shift_logits.size(-1)), shift_labels.view(-1))
+        # The up-cast to float32 is deliberate and must not be removed. Without it the log-softmax,
+        # its backward, and the mean reduction all run in the logits' own dtype -- and under mixed
+        # precision the model hands us bfloat16. FSDP2's MixedPrecisionPolicy casts *parameters*; it
+        # does not install torch.autocast, so nothing else promotes them. TorchTitan up-casts on
+        # every one of its cross-entropy paths for the same reason (torchtitan/components/loss.py),
+        # as does HF transformers (transformers/loss/loss_utils.py).
+        loss = self.loss_fun(shift_logits.view(-1, shift_logits.size(-1)).float(), shift_labels.view(-1))
         return loss
 
     def _parse_arguments(
